requirements.txt
notebook
pandas
numpy
seaborn
matplotlib
missingno
psycopg2-binary
python-dotenv